Move event handler tests to separate file
