+++ pyproject.toml
@@ -65,6 +65,7 @@ exclude_also = [
     "if settings.DEBUG",
     "raise AssertionError",
     "raise NotImplementedError",
+    "pass",
     "if 0:",
     "if __name__ == .__main__.:",
     "if TYPE_CHECKING:",
