+++ tests/test_event_bus.py
@@ -23,8 +23,7 @@ class ExplosionEvent(Event):
 
 
 class FloodEvent:
-    def print_event_data(self) -> None:
-        print("It got wet")
+    pass
 
 
 class AnyEventHandler(IMessageHandler):
@@ -41,48 +40,6 @@ class SecondExplosionEventHandler(IMessageHandler):
     def handle(self, event: ExplosionEvent) -> None:
         event.print_event_data()
         print("again")
-
-
-class TestEventHandler:
-    def test_non_specific_event_handler_can_handle_an_event(
-        self, capsys: CaptureFixture[str]
-    ) -> None:
-        event = ExplosionEvent()
-        handler = AnyEventHandler()
-
-        handler.handle(event)
-
-        captured = capsys.readouterr()
-        assert captured.out == "Hi\n"
-
-    def test_specific_event_handler_can_handle_a_valid_event(
-        self, capsys: CaptureFixture[str]
-    ) -> None:
-        event = ExplosionEvent()
-        handler = ExplosionEventHandler()
-
-        handler.handle(event)
-
-        captured = capsys.readouterr()
-        assert captured.out == "It went boom\n"
-
-    def test_event_handler_does_not_restrict_event_type(
-        self, capsys: CaptureFixture[str]
-    ) -> None:
-        """Remove responsibility from the handlers and allow duck typing.
-
-        This will not pass static type checking but the event type is not enforced
-        by handlers at runtime. Specific handlers are created by users, so it's
-        difficult to enforce type checking. Instead, this responsibility
-        is handled by the Event Bus.
-        """
-        event = FloodEvent()
-        handler = ExplosionEventHandler()
-
-        handler.handle(event)  # type: ignore
-
-        captured = capsys.readouterr()
-        assert captured.out == "It got wet\n"
 
 
 class TestEventBus:
